+++ NESseract/NESseract.Core/Cpu/Operations/PLAOperation.cs
@@ -10,7 +10,7 @@ namespace NESseract.Core.Cpu.Operations
          var operationValue = memory.Stack.Span[++registers.SP];
 
          registers.N_NegativeFlag = (byte)((operationValue & 0x80) >> 7);
-         registers.Z_ZeroFlag = operationValue == 0 ? 1 : 0;
+         registers.Z_ZeroFlag = (byte)(operationValue == 0 ? 1 : 0);
 
          registers.A = operationValue;
